requirements.txt
python-telegram-bot==13.1
spotipy==2.23.0
deep-translator==1.10.1
pytz==2023.3
